requirements.txt
rasterio~=1.2.6
Pillow~=8.3.1
numpy~=1.21.1
scipy~=1.7.0
imageio~=2.9.0

